Refactored list pages.
