+++ src/app/pages/lists/organizations/index.tsx
@@ -1,9 +1,15 @@
 import React, { useState, useEffect } from 'react';
 
-import { Container, Loader } from 'semantic-ui-react';
+import { Container, Loader, Header } from 'semantic-ui-react';
 
-import ListView from './listView';
+import { faUsers } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+import ListView from '../listView';
+import listItemView from './listItemView';
 // import DetailView from './detailView';
+
+import localStyles from './style.less';
 
 const dataSourceUrl = 'https://api.acikkaynak.info/lists/organizations';
 
@@ -52,7 +58,19 @@ function Organizations() {
 
     return (
         <Container className="content">
-            <ListView datasource={organizations} />
+            <Header as="h1" className={localStyles.orgHeader}>
+                <i aria-hidden="true" className="circular icon">
+                    <FontAwesomeIcon icon={faUsers} />
+                </i>
+                <Header.Content>
+                    Organizasyonlar
+                    <Header.Subheader>
+                        Organizasyonlar Listesi
+                    </Header.Subheader>
+                </Header.Content>
+            </Header>
+
+            <ListView datasource={organizations} listItemView={listItemView} />
         </Container>
     );
 }
