+++ src/app/pages/lists/style.less
@@ -1,0 +1,19 @@
+.ui.positive.button.category-link {
+  margin: 0.3em 0.3em 0 0;
+}
+
+.ui.segment.category-segment {
+  word-break: break-word;
+}
+
+@media only screen and (max-width: 600px) {
+  .projects {
+      .content {
+          font-size: 28px;
+      }
+
+      .sub.header {
+          font-size: 20px !important;
+      }
+  }
+}
